frontend roles in interface

+++ XYZEngineeringProject.UI/src/app/components/settings/department/department-index/department-index.component.ts
@@ -1,9 +1,12 @@
+import { TranslateService } from '@ngx-translate/core';
 import { DepartmentManager } from './../../../../models/department.model';
 import { Department } from 'src/app/models/department.model';
 import { User } from 'src/app/models/user.model';
 import { UserService } from './../../../../services/user/user.service';
 import { DepartmentService } from './../../../../services/department/department.service';
 import { Component, OnInit } from '@angular/core';
+import { RolesDB } from 'src/app/models/roles.enum';
+import { AuthorizationService } from 'src/app/services/authorization/authorization.service';
 
 @Component({
   selector: 'app-department-index',
@@ -13,17 +16,29 @@ import { Component, OnInit } from '@angular/core';
 export class DepartmentIndexComponent implements OnInit {
 
   departments: Department[] = [];
-
   departmentsMan: DepartmentManager[] = [];
-
-  managers: User[] = []
+  managers: User[] = [];
+  currentUser: User;
+  canModifyRoles: RolesDB[] = [
+    RolesDB.Admin,
+    RolesDB.Moderator,
+    RolesDB.Management
+  ];
 
   constructor(
     private departmentService: DepartmentService,
-    private userService: UserService
+    private userService: UserService,
+    private translateService: TranslateService,
+    private authorizationService: AuthorizationService
   ) {}
 
   ngOnInit(): void {
+    this.authorizationService.currentUser().subscribe({
+      next: (res) => {
+        this.currentUser = res;
+      }
+    });
+
     this.departmentService.getAllDepartments().subscribe({
       next: (departmentsResponse) => {
         this.departments = departmentsResponse;
@@ -52,16 +67,26 @@ export class DepartmentIndexComponent implements OnInit {
     })
   }
 
-  canDelete(department:Department){
-    return true
+  deleteDepartment(department: Department) {
+    if (confirm(this.translateService.instant('Alert.deleteDepartment') + department.name + "?")) {
+      this.departmentService.deleteDepartment(department).subscribe({
+        next: (res) => {
+          window.location.reload();
+        }
+      })
+    }
   }
-
-  deleteDepartment(department:Department){
-    this.departmentService.deleteDepartment(department).subscribe({
-      next: (res) => {
-        window.location.reload();
+  
+  canModify() {
+    let canModify: boolean = false;
+    
+    this.canModifyRoles.forEach(role => {
+      if (this.currentUser.roles.includes(role)) {
+        canModify = true;
       }
-    })
+    });
+   
+    return canModify;
   }
 
 }
